separate edge case tests
